+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 1.3.4
+# 1.3.4 - 3 Jun 2025
 Feature:
 - sse helper